refactor submission form into its own file
